+++ get_test.go
@@ -82,6 +82,8 @@ func TestGet(t *testing.T) {
 			//}
 			//fmt.Println(res.Message, res.ResCode)
 			value, _ := res.Value.(GetResultType)
+			fmt.Printf("get-by-id-value: %#v\n", value.TableRecords)
+			fmt.Printf("get-by-param-count: %v\n", value.RecordCount)
 			jsonRecs, _ := json.Marshal(value.TableRecords)
 			fmt.Printf("json-records: %v\n\n", string(jsonRecs))
 			mctest.AssertEquals(t, res.Code, "success", "get-task should return code: success")
@@ -103,8 +105,10 @@ func TestGet(t *testing.T) {
 			)
 			tableFieldPointers := []interface{}{&id, &tableName, &logRecords, &newLogRecords, &logBy, &logType, &logAt}
 			res := getCrud.GetByParam(GetTableFields, tableFieldPointers)
-			fmt.Printf("get-by-param-response: %#v\n", res)
+			//fmt.Printf("get-by-param-response: %#v\n", res)
 			value, _ := res.Value.(GetResultType)
+			fmt.Printf("get-by-param-value: %#v\n", value.TableRecords)
+			fmt.Printf("get-by-param-count: %v\n", value.RecordCount)
 			mctest.AssertEquals(t, res.Code, "success", "get-task should return code: success")
 			mctest.AssertEquals(t, value.RecordCount >= 0, true, "get-task-count should be >= 0")
 			mctest.AssertEquals(t, len(value.TableRecords) >= 0, true, "get-result-count should be >= 0")
@@ -128,6 +132,7 @@ func TestGet(t *testing.T) {
 			value, _ := res.Value.(GetResultType)
 			fmt.Printf("get-by-all-value[0]: %#v\n", value.TableRecords[0])
 			fmt.Printf("get-by-all-value[1]: %#v\n", value.TableRecords[1])
+			fmt.Printf("get-by-all-count: %v\n", value.RecordCount)
 			mctest.AssertEquals(t, res.Code, "success", "get-task should return code: success")
 			mctest.AssertEquals(t, value.RecordCount >= 10, true, "get-task-count should be >= 10")
 			mctest.AssertEquals(t, len(value.TableRecords) >= 10, true, "get-result-count should be >= 10")
@@ -152,6 +157,7 @@ func TestGet(t *testing.T) {
 			value, _ := res.Value.(GetResultType)
 			fmt.Printf("get-by-all-value[0]: %#v\n", value.TableRecords[0])
 			fmt.Printf("get-by-all-value[1]: %#v\n", value.TableRecords[1])
+			fmt.Printf("get-by-all-limit-count: %v\n", value.RecordCount)
 			mctest.AssertEquals(t, res.Code, "success", "get-task should return code: success")
 			mctest.AssertEquals(t, value.RecordCount == 20, true, "get-task-count should be = 20")
 			mctest.AssertEquals(t, len(value.TableRecords) == 20, true, "get-result-count should be = 20")
@@ -161,4 +167,3 @@ func TestGet(t *testing.T) {
 	mctest.PostTestResult()
 
 }
-